pubgrub

Enum DerivationTree

pub enum DerivationTree<P: Package, VS: VersionSet, M: Eq + Clone + Debug + Display> {
    External(External<P, VS, M>),
    Derived(Derived<P, VS, M>),
}

Derivation tree resulting in the impossibility to solve the dependencies of our root package.

Variants

External(External<P, VS, M>)

External incompatibility.

Derived(Derived<P, VS, M>)

Incompatibility derived from two others.

Implementations

impl<P: Package, VS: VersionSet, M: Eq + Clone + Debug + Display> DerivationTree<P, VS, M>

pub fn packages(&self) -> Set<&P>

Get all packages referred to in the derivation tree.

pub fn collapse_no_versions(&mut self)

Merge the NoVersions external incompatibilities with the other one they are matched with in a derived incompatibility. This cleans up quite nicely the generated report. You might want to do this if you know that the DependencyProvider was not run in some kind of offline mode that may not have access to all versions existing.