Enum pubgrub::error::PubGrubError

pub enum PubGrubError<DP>
where
    DP: DependencyProvider,
{
    NoSolution(DerivationTree<DP::P, DP::VS>),
    ErrorRetrievingDependencies {
        package: DP::P,
        version: DP::V,
        source: DP::Err,
    },
    SelfDependency {
        package: DP::P,
        version: DP::V,
    },
    ErrorChoosingPackageVersion(DP::Err),
    ErrorInShouldCancel(DP::Err),
    Failure(String),
}

Errors that may occur while solving dependencies.

Variants

NoSolution(DerivationTree<DP::P, DP::VS>)

There is no solution for this set of dependencies.

ErrorRetrievingDependencies

Error arising when the implementer of DependencyProvider returned an error in the method get_dependencies.

Fields

package: DP::P

Package whose dependencies we want.

version: DP::V

Version of the package for which we want the dependencies.

source: DP::Err

Error raised by the implementer of DependencyProvider.

SelfDependency

Error arising when the implementer of DependencyProvider returned a dependency on the requested package. This technically means that the package directly depends on itself, and is clearly some kind of mistake.

Fields

package: DP::P

Package whose dependencies we want.

version: DP::V

Version of the package for which we want the dependencies.

ErrorChoosingPackageVersion(DP::Err)

Error arising when the implementer of DependencyProvider returned an error in the method choose_version.

ErrorInShouldCancel(DP::Err)

Error arising when the implementer of DependencyProvider returned an error in the method should_cancel.

Failure(String)

Something unexpected happened.