Module pubgrub::report

Build a report as clear as possible as to why dependency solving failed.

Structs

DefaultStringReporter

Default reporter able to generate an explanation as a String.

Derived

Incompatibility derived from two others.

Enums

DerivationTree

Derivation tree resulting in the impossibility to solve the dependencies of our root package.

External

Incompatibilities that are not derived from others, they have their own reason.

Traits

Reporter

Reporter trait.