Trait pubgrub::report::Reporter

pub trait Reporter<P: Package, VS: VersionSet> {
    type Output;

    // Required methods
    fn report(derivation_tree: &DerivationTree<P, VS>) -> Self::Output;
    fn report_with_formatter(
        derivation_tree: &DerivationTree<P, VS>,
        formatter: &impl ReportFormatter<P, VS, Output = Self::Output>
    ) -> Self::Output;
}

Reporter trait.

Required Associated Types

type Output

Output type of the report.

Required Methods

fn report(derivation_tree: &DerivationTree<P, VS>) -> Self::Output

Generate a report from the derivation tree describing the resolution failure using the default formatter.

fn report_with_formatter( derivation_tree: &DerivationTree<P, VS>, formatter: &impl ReportFormatter<P, VS, Output = Self::Output> ) -> Self::Output

Generate a report from the derivation tree describing the resolution failure using a custom formatter.

Object Safety

This trait is not object safe.