Struct pubgrub::solver::OfflineDependencyProvider

pub struct OfflineDependencyProvider<P: Package, V: Version> { /* fields omitted */ }

A basic implementation of DependencyProvider.

Implementations

impl<P: Package, V: Version> OfflineDependencyProvider<P, V>

pub fn new() -> Self

Creates an empty OfflineDependencyProvider with no dependencies.

pub fn add_dependencies<I: IntoIterator<Item = (P, Range<V>)>>(
&mut self,
package: P,
version: impl Into<V>,
dependencies: I
)

Registers the dependencies of a package and version pair. Dependencies must be added with a single call to add_dependencies. All subsequent calls to add_dependencies for a given package version pair will replace the dependencies by the new ones.

The API does not allow to add dependencies one at a time to uphold an assumption that OfflineDependencyProvider.get_dependencies(p, v) provides all dependencies of a given package (p) and version (v) pair.

pub fn packages(&self) -> impl Iterator<Item = &P>

Lists packages that have bean saved.

pub fn versions(&self, package: &P) -> Option<impl Iterator<Item = &V>>

Lists versions of saved packages in sorted order. Returns None if no information is available regarding that package.