pubgrub

Trait Reporter

pub trait Reporter<P: Package, VS: VersionSet, M: Eq + Clone + Debug + Display> {
    type Output;

    // Required methods
    fn report(derivation_tree: &DerivationTree<P, VS, M>) -> Self::Output;
    fn report_with_formatter(
        derivation_tree: &DerivationTree<P, VS, M>,
        formatter: &impl ReportFormatter<P, VS, M, Output = Self::Output>,
    ) -> Self::Output;
}

Reporter trait.

Required Associated Types

type Output

Output type of the report.

Required Methods

fn report(derivation_tree: &DerivationTree<P, VS, M>) -> Self::Output

Generate a report from the derivation tree describing the resolution failure using the default formatter.

fn report_with_formatter( derivation_tree: &DerivationTree<P, VS, M>, formatter: &impl ReportFormatter<P, VS, M, Output = Self::Output>, ) -> Self::Output

Generate a report from the derivation tree describing the resolution failure using a custom formatter.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".