pubgrub

Trait VersionSet

pub trait VersionSet:
    Debug
    + Display
    + Clone
    + Eq {
    type V: Debug + Display + Clone + Ord;

    // Required methods
    fn empty() -> Self;
    fn singleton(v: Self::V) -> Self;
    fn complement(&self) -> Self;
    fn intersection(&self, other: &Self) -> Self;
    fn contains(&self, v: &Self::V) -> bool;

    // Provided methods
    fn contains_many<'s, I, BV>(
        &'s self,
        versions: I,
    ) -> impl Iterator<Item = bool> + 's
       where I: Iterator<Item = BV> + 's,
             BV: Borrow<Self::V> + 's { ... }
    fn full() -> Self { ... }
    fn union(&self, other: &Self) -> Self { ... }
    fn difference(&self, other: &Self) -> Self { ... }
    fn is_disjoint(&self, other: &Self) -> bool { ... }
    fn subset_of(&self, other: &Self) -> bool { ... }
    fn relation(&self, other: &Self) -> SetRelation { ... }
}

A set of versions.

See Ranges for an implementation.

The methods with default implementations can be overwritten for better performance, but their output must be equal to the default implementation.

Equality

It is important that the Eq trait is implemented so that if two sets contain the same versions, they are equal under Eq. In particular, you can only use #[derive(Eq, PartialEq)] if Eq is strictly equivalent to the structural equality, i.e. if version sets are always stored in a canonical representations. Such problems may arise if your implementations of complement() and intersection() do not return canonical representations.

For example, >=1,<4 || >=2,<5 and >=1,<4 || >=3,<5 are equal, because they can both be normalized to >=1,<5.

Note that pubgrub does not know which versions actually exist for a package, the contract is about upholding the mathematical properties of set operations, assuming all versions are possible. This is required for the solver to determine the relationship of version sets to each other.

Required Associated Types

type V: Debug + Display + Clone + Ord

Version type associated with the sets manipulated.

Required Methods

fn empty() -> Self

An empty set containing no version.

fn singleton(v: Self::V) -> Self

A set containing only the given version.

fn complement(&self) -> Self

The set of all version that are not in this set.

fn intersection(&self, other: &Self) -> Self

The set of all versions that are in both sets.

fn contains(&self, v: &Self::V) -> bool

Whether the version is part of this set.

Provided Methods

fn contains_many<'s, I, BV>( &'s self, versions: I, ) -> impl Iterator<Item = bool> + 's
where I: Iterator<Item = BV> + 's, BV: Borrow<Self::V> + 's,

Whether each version is part of this set, in iterator order.

versions must be in nondecreasing order. Implementations can override this to walk the versions and the set together.

fn full() -> Self

The set containing all versions.

The default implementation is the complement of the empty set.

fn union(&self, other: &Self) -> Self

The set of all versions that are either (or both) of the sets.

The default implementation is complement of the intersection of the complements of both sets (De Morgan’s law).

fn difference(&self, other: &Self) -> Self

The set of versions that are in self but not in other.

Implementations can override this to avoid materializing the complement.

fn is_disjoint(&self, other: &Self) -> bool

Whether the ranges have no overlapping segments.

fn subset_of(&self, other: &Self) -> bool

Whether all ranges of self are contained in other.

fn relation(&self, other: &Self) -> SetRelation

Classifies self as a subset of, disjoint from, or partially overlapping with other.

Implementations can override this to avoid traversing both sets once for Self::subset_of and again for Self::is_disjoint. An empty self must be classified as SetRelation::Subset.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".