Trait pubgrub::version::Version

pub trait Version: Clone + Ord + Debug + Display {
    // Required methods
    fn lowest() -> Self;
    fn bump(&self) -> Self;
}

Versions have a minimal version (a “0” version) and are ordered such that every version has a next one.

Required Methods

fn lowest() -> Self

Returns the lowest version.

fn bump(&self) -> Self

Returns the next version, the smallest strictly higher version.

Object Safety

This trait is not object safe.