Trait pubgrub::version_set::VersionSet

pub trait VersionSet: Debug + Display + Clone + Eq {
    type V: Debug + Display + Clone + Ord;

    // Required methods
    fn empty() -> Self;
    fn singleton(v: Self::V) -> Self;
    fn complement(&self) -> Self;
    fn intersection(&self, other: &Self) -> Self;
    fn contains(&self, v: &Self::V) -> bool;

    // Provided methods
    fn full() -> Self { ... }
    fn union(&self, other: &Self) -> Self { ... }
    fn is_disjoint(&self, other: &Self) -> bool { ... }
    fn subset_of(&self, other: &Self) -> bool { ... }
}

Trait describing sets of versions.

Required Associated Types

type V: Debug + Display + Clone + Ord

Version type associated with the sets manipulated.

Required Methods

fn empty() -> Self

Constructor for an empty set containing no version.

fn singleton(v: Self::V) -> Self

Constructor for a set containing exactly one version.

fn complement(&self) -> Self

Compute the complement of this set.

fn intersection(&self, other: &Self) -> Self

Compute the intersection with another set.

fn contains(&self, v: &Self::V) -> bool

Evaluate membership of a version in this set.

Provided Methods

fn full() -> Self

Constructor for the set containing all versions. Automatically implemented as Self::empty().complement().

fn union(&self, other: &Self) -> Self

Compute the union with another set. Thanks to set properties, this is automatically implemented as: self.complement().intersection(&other.complement()).complement()

fn is_disjoint(&self, other: &Self) -> bool

Whether the range have no overlapping segmets

fn subset_of(&self, other: &Self) -> bool

Whether all range of self are contained in other

Object Safety

This trait is not object safe.