Struct pubgrub::solver::OfflineDependencyProvider

pub struct OfflineDependencyProvider<P: Package, VS: VersionSet> { /* private fields */ }

A basic implementation of DependencyProvider.

Implementations

impl<P: Package, VS: VersionSet> OfflineDependencyProvider<P, VS>

pub fn new() -> Self

Creates an empty OfflineDependencyProvider with no dependencies.

pub fn add_dependencies<I: IntoIterator<Item = (P, VS)>>( &mut self, package: P, version: impl Into<VS::V>, dependencies: I )

Registers the dependencies of a package and version pair. Dependencies must be added with a single call to add_dependencies. All subsequent calls to add_dependencies for a given package version pair will replace the dependencies by the new ones.

The API does not allow to add dependencies one at a time to uphold an assumption that OfflineDependencyProvider.get_dependencies(p, v) provides all dependencies of a given package (p) and version (v) pair.

pub fn packages(&self) -> impl Iterator<Item = &P>

Lists packages that have been saved.

pub fn versions(&self, package: &P) -> Option<impl Iterator<Item = &VS::V>>

Lists versions of saved packages in sorted order. Returns None if no information is available regarding that package.